requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.1.1
idna==3.7
iniconfig==2.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
playwright==1.49.0
pluggy==1.5.0
pycparser==2.22
pyee==12.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
pyTelegramBotAPI==4.20.0
pytest==8.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
telebot==0.0.5
text-unidecode==1.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
